ENH: Move traceback formatter to from tests.utils to helpers
Now allows for tests to be removed from package if desired
